Complete refocus of project
Previously this project behaved as a sort of documentation. However,
Technical Machine made many changes to their Tessel prototype. To
prevent having to constantly update this project, I've decided to remove
the documentation aspect of this project and instead focus solely on the
keyboard functionality. Besides, their documentation has improved
substantially. This commit removes the documentation-like code and lays
the ground work for the keyboard "admin."